Refactor codebook and icdf
