+++ symphonia-codec-opus/src/silk/mod.rs
@@ -6,8 +6,7 @@ mod table;
 mod codebook;
 mod error;
 mod decoder;
-mod icdf;
 
-pub (crate) use decoder::Decoder;
+pub(crate) use decoder::Decoder;
 pub use error::Error;
 pub use frame::{Frame, FrameType, QuantizationOffsetType};
